[#48] refactor: compound component pattern으로 변경
