+++ components/common/Form/Form.module.scss
@@ -1,0 +1,2 @@
+.wrapper {
+}
