+++ components/common/Input/Input.module.scss
@@ -1,12 +1,11 @@
 .container {
   display: flex;
   flex-direction: column;
-  gap: 120px;
-  margin-bottom: 41px;
+  gap: 100px;
 }
 
 .input {
-  outline: none;
+  // outline: none;
   border-width: 0 0 1.5px;
   border-color: var(--main-color);
 }
